requirements.txt
requests==2.31.0
ics==0.7.2
zhconv==1.4.3
python-dateutil==2.8.2